Implements enemy moving, hud, bullet deletion

+++ Assets/Scripts/Entity.cs
@@ -3,9 +3,9 @@ using System.Collections;
 
 public class Entity : MonoBehaviour {
 
-	public float health;
+	public int health;
 
-	public void TakeDamage(float damage) {
+	public void TakeDamage(int damage) {
 		health -= damage;
 
 		if (health <= 0) {
